Add tenant support for Session Cookie verification
It doesn't seem to be supported at the moment (executing it with a
tenant-enabled Firebase project yields an
`UNSUPPORTED_TENANT_OPERATION`) error, but once it _is_ supported,
this library will need no or just minimal updates.

The Firebase Admin SDK for PHP has integration tests checking for this
error so that we know early on when it starts working.